This is the 4.14.74 stable release


Conflicts:
	arch/arm64/kvm/guest.c
	drivers/input/misc/xen-kbdfront.c
	include/linux/slub_def.h

Changes to be committed:
 	modified:   Documentation/hwmon/ina2xx
 	modified:   arch/arm/boot/dts/dra7.dtsi
 	modified:   arch/arm/boot/dts/imx7d.dtsi
 	modified:   arch/arm/boot/dts/ls1021a.dtsi
 	modified:   arch/arm/boot/dts/mt7623.dtsi
 	modified:   arch/arm/boot/dts/omap4-droid4-xt894.dts
 	modified:   arch/arm/mach-mvebu/pmsu.c
 	modified:   arch/arm/mach-omap2/omap_hwmod_reset.c
 	modified:   arch/mips/boot/Makefile
 	modified:   arch/powerpc/kernel/machine_kexec.c
 	modified:   arch/powerpc/platforms/powernv/pci-ioda.c
 	modified:   arch/s390/kernel/sysinfo.c
 	modified:   arch/s390/mm/extmem.c
 	modified:   arch/s390/mm/pgalloc.c
 	modified:   arch/x86/entry/entry_64.S
 	modified:   arch/x86/events/intel/lbr.c
 	modified:   arch/x86/events/perf_event.h
 	modified:   arch/x86/include/asm/fixmap.h
 	modified:   arch/x86/include/asm/pgtable_64.h
 	modified:   arch/x86/kernel/head64.c
 	modified:   arch/x86/kernel/head_64.S
 	modified:   arch/x86/kernel/tsc_msr.c
 	modified:   arch/x86/mm/numa_emulation.c
 	modified:   arch/x86/mm/pgtable.c
 	modified:   arch/x86/mm/pti.c
 	modified:   arch/x86/xen/mmu_pv.c
 	modified:   crypto/ablkcipher.c
 	modified:   crypto/blkcipher.c
 	modified:   drivers/ata/pata_ftide010.c
 	modified:   drivers/bluetooth/btusb.c
 	modified:   drivers/crypto/cavium/nitrox/nitrox_dev.h
 	modified:   drivers/crypto/cavium/nitrox/nitrox_lib.c
 	modified:   drivers/crypto/cavium/nitrox/nitrox_reqmgr.c
 	modified:   drivers/edac/edac_mc_sysfs.c
 	modified:   drivers/edac/i7core_edac.c
 	modified:   drivers/gpio/gpio-menz127.c
 	modified:   drivers/gpu/drm/amd/amdgpu/gfx_v8_0.c
 	modified:   drivers/gpu/drm/amd/amdgpu/kv_dpm.c
 	modified:   drivers/gpu/drm/amd/amdgpu/si_dpm.c
 	modified:   drivers/gpu/drm/i915/i915_gem.c
 	modified:   drivers/gpu/drm/i915/i915_vma.c
 	modified:   drivers/gpu/drm/sun4i/sun4i_drv.c
 	modified:   drivers/hid/hid-ntrig.c
 	modified:   drivers/hwmon/adt7475.c
 	modified:   drivers/hwmon/ina2xx.c
 	modified:   drivers/hwtracing/intel_th/core.c
 	modified:   drivers/i2c/busses/i2c-i801.c
 	modified:   drivers/iio/accel/adxl345_core.c
 	modified:   drivers/iio/adc/ina2xx-adc.c
 	modified:   drivers/iio/counter/104-quad-8.c
 	modified:   drivers/infiniband/core/rw.c
 	modified:   drivers/infiniband/core/uverbs_main.c
 	modified:   drivers/infiniband/hw/bnxt_re/qplib_sp.c
 	modified:   drivers/infiniband/hw/hfi1/pio.c
 	modified:   drivers/infiniband/hw/hfi1/user_sdma.c
 	modified:   drivers/infiniband/hw/hfi1/verbs.c
 	modified:   drivers/infiniband/hw/i40iw/i40iw_verbs.c
 	modified:   drivers/infiniband/hw/mlx4/qp.c
 	modified:   drivers/infiniband/ulp/srp/ib_srp.c
 	modified:   drivers/input/mouse/elantech.c
 	modified:   drivers/iommu/amd_iommu.c
 	modified:   drivers/iommu/msm_iommu.c
 	modified:   drivers/md/md-cluster.c
 	modified:   drivers/media/i2c/soc_camera/ov772x.c
 	modified:   drivers/media/platform/exynos4-is/fimc-isp-video.c
 	modified:   drivers/media/platform/fsl-viu.c
 	modified:   drivers/media/platform/omap3isp/isp.c
 	modified:   drivers/media/platform/s3c-camif/camif-capture.c
 	modified:   drivers/media/usb/tm6000/tm6000-dvb.c
 	modified:   drivers/media/usb/uvc/uvc_video.c
 	modified:   drivers/media/v4l2-core/v4l2-event.c
 	modified:   drivers/media/v4l2-core/v4l2-fh.c
 	modified:   drivers/misc/sram.c
 	modified:   drivers/misc/tsl2550.c
 	modified:   drivers/misc/vmw_vmci/vmci_queue_pair.c
 	modified:   drivers/mtd/nand/atmel/nand-controller.c
 	modified:   drivers/net/ethernet/hisilicon/hns/hnae.h
 	modified:   drivers/net/ethernet/hisilicon/hns/hns_enet.c
 	modified:   drivers/net/ethernet/intel/e1000/e1000_ethtool.c
 	modified:   drivers/net/ethernet/qlogic/qed/qed_mcp.c
 	modified:   drivers/net/ethernet/qlogic/qed/qed_mcp.h
 	modified:   drivers/net/ethernet/qlogic/qed/qed_reg_addr.h
 	modified:   drivers/net/phy/xilinx_gmii2rgmii.c
 	modified:   drivers/net/wireless/ath/ath10k/htt_rx.c
 	modified:   drivers/net/wireless/ath/ath10k/mac.c
 	modified:   drivers/net/wireless/ath/ath10k/sdio.c
 	modified:   drivers/net/wireless/broadcom/brcm80211/brcmsmac/phy/phy_qmath.c
 	modified:   drivers/net/wireless/rndis_wlan.c
 	modified:   drivers/net/wireless/ti/wlcore/cmd.c
 	modified:   drivers/nvme/target/fcloop.c
 	modified:   drivers/power/reset/vexpress-poweroff.c
 	modified:   drivers/power/supply/axp288_charger.c
 	modified:   drivers/power/supply/power_supply_core.c
 	modified:   drivers/regulator/core.c
 	modified:   drivers/s390/block/dasd.c
 	modified:   drivers/s390/block/scm_blk.c
 	modified:   drivers/scsi/bnx2i/bnx2i_hwi.c
 	modified:   drivers/scsi/ibmvscsi/ibmvscsi.c
 	modified:   drivers/scsi/megaraid/megaraid_sas_base.c
 	modified:   drivers/spi/spi-rspi.c
 	modified:   drivers/spi/spi-sh-msiof.c
 	modified:   drivers/spi/spi-tegra20-slink.c
 	modified:   drivers/staging/android/ashmem.c
 	modified:   drivers/staging/media/imx/imx-ic-prpencvf.c
 	modified:   drivers/staging/media/imx/imx-media-csi.c
 	modified:   drivers/staging/rts5208/sd.c
 	modified:   drivers/target/iscsi/iscsi_target_tpg.c
 	modified:   drivers/target/target_core_device.c
 	modified:   drivers/thermal/of-thermal.c
 	modified:   drivers/tty/serial/8250/serial_cs.c
 	modified:   drivers/tty/serial/cpm_uart/cpm_uart_core.c
 	modified:   drivers/tty/serial/fsl_lpuart.c
 	modified:   drivers/tty/serial/imx.c
 	modified:   drivers/tty/serial/sh-sci.c
 	modified:   drivers/usb/class/cdc-wdm.c
 	modified:   drivers/usb/core/devio.c
 	modified:   drivers/usb/core/driver.c
 	modified:   drivers/usb/core/usb.c
 	modified:   drivers/usb/musb/musb_dsps.c
 	modified:   drivers/usb/serial/kobil_sct.c
 	modified:   drivers/usb/wusbcore/security.c
 	modified:   drivers/uwb/hwa-rc.c
 	modified:   fs/iomap.c
 	modified:   fs/isofs/inode.c
 	modified:   fs/locks.c
 	modified:   fs/nfsd/nfs4proc.c
 	modified:   fs/overlayfs/inode.c
 	modified:   include/linux/arm-smccc.h
 	modified:   include/linux/platform_data/ina2xx.h
 	modified:   include/linux/power_supply.h
 	modified:   include/linux/slub_def.h
 	modified:   include/media/v4l2-fh.h
 	modified:   kernel/bpf/sockmap.c
 	modified:   kernel/module.c
 	modified:   kernel/time/alarmtimer.c
 	modified:   kernel/time/posix-timers.c
 	modified:   kernel/time/posix-timers.h
 	modified:   lib/klist.c
 	modified:   mm/slub.c
 	modified:   net/6lowpan/iphc.c
 	modified:   net/tls/tls_main.c
 	modified:   sound/aoa/core/gpio-feature.c
 	modified:   sound/pci/hda/hda_intel.c
 	modified:   sound/soc/soc-dapm.c

+++ kernel/time/posix-timers.c
@@ -654,11 +654,11 @@ static ktime_t common_hrtimer_remaining(struct k_itimer *timr, ktime_t now)
 	return __hrtimer_expires_remaining_adjusted(timer, now);
 }
 
-static int common_hrtimer_forward(struct k_itimer *timr, ktime_t now)
+static s64 common_hrtimer_forward(struct k_itimer *timr, ktime_t now)
 {
 	struct hrtimer *timer = &timr->it.real.timer;
 
-	return (int)hrtimer_forward(timer, now, timr->it_interval);
+	return hrtimer_forward(timer, now, timr->it_interval);
 }
 
 /*
